fix: resolve issue where app crashes when user navigates to the meal tab.
The reason for the crash is the structure of the response change to accomodate pagination and follow conventions of the other responses. Changing the structure fixes the crashing issues.

Pagination was also factored in the fix. The next 20 meals will be fetched and appended to the existing list when the user scrolls to the bottom of the page.